+++ src/Exceptions/RateLimitExceededException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the RateLimitExceededException class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-31
+ *
+ */
+
+namespace VaniloCloud\Exceptions;
+
+use RuntimeException;
+
+class RateLimitExceededException extends RuntimeException
+{
+}