+++ Changelog.md
@@ -1,5 +1,10 @@
 # Vanilo Cloud PHP SDK Changelog
 
+## Unreleased
+##### 2023-XX-YY
+
+- Fixed Taxonomy hydration in its list endpoint
+
 ## 0.1.0
 ##### 2023-01-31